fix: checkbox listener in BookingActivity

+++ app/src/main/java/by/story_weaver/ridereserve/Logic/data/repositories/impl/UserRepositoryImpl.java
@@ -61,7 +61,7 @@ public class UserRepositoryImpl implements UserRepository {
         dao.setUserInSystem(id);
     }
     @Override
-    public int getUserInSystem(){
+    public long getUserInSystem(){
         return dao.getIdUserInSystem();
     }
 }
