+++ app/src/main/java/by/story_weaver/ridereserve/Logic/viewModels/AuthViewModel.java
@@ -34,8 +34,8 @@ public class AuthViewModel extends ViewModel {
     public LiveData<UiState<User>> getUserStateEnter() { return UserStateEnter; }
     public LiveData<UiState<User>> getUserStateReg() { return UserStateReg; }
 
-    public int checkSignedIn() {
-        int id = userRepo.getUserInSystem();
+    public long checkSignedIn() {
+        long id = userRepo.getUserInSystem();
         if (id > 0) {
             return id;
         } else {
